requirements.txt
python-dotenv
igdb-api-v4